requirements.txt
Flask
repoze.lru
sh
youtube_dl
